moved product import to tmp
